@@ -3,21 +3,23 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "bootstrap": "^4.1.3",
+    "@material-ui/core": "^3.9.3",
     "chart.js": "^2.7.2",
     "dexie": "^2.0.4",
     "es6-promise": "^4.2.6",
-    "font-awesome": "^4.7.0",
-    "jquery": "^3.3.1",
+    "firebase": "^5.10.1",
     "lodash": "^4.17.10",
     "luxon": "^1.3.3",
     "node-sass-chokidar": "^1.3.4",
     "popper.js": "^1.14.4",
     "rcolor": "^1.0.1",
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0",
+    "react-redux": "^7.0.3",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.5"
+    "react-scripts": "2.1.5",
+    "redux": "^4.0.1",
+    "redux-thunk": "^2.3.0"
   },
   "scripts": {
     "build-css": "node-sass-chokidar src/ -o src/",
