@@ -1,114 +1,107 @@
 import React, { Component } from 'react';
-import { budget, transactions } from '../../../db/db';
+import { connect } from 'react-redux';
+import { withRouter } from 'react-router-dom';
 import { DateTime } from 'luxon';
 
-class AddExpense extends Component {
-    constructor(props) {
-        super(props);
-        this.today = DateTime.local();
-        this.state = {
-            expense: {
-                desc: '',
-                amount: '',
-                date: this.today.toISODate(),
-                budgetId: ''
-            },
-            budgetList: []
-        };
-        this.updateProperty = this.updateProperty.bind(this);
-        this.addNewExpense = this.addNewExpense.bind(this);
-        this.goBack = this.goBack.bind(this);
-    }
+import Amount from '../../common/amount/Amount';
+import Category from "../../common/category/Category";
+import Description from '../../common/description/Description';
 
-    fetchExpenseBudget() {
-        budget.where('type').equals('EXPENSE')
-            .filter((budget) => {
-                return ((budget.year === this.today.year &&
-                    budget.month === this.today.month + 1) || budget.isRecurring);
-            })
-            .toArray()
-            .then((response) => {
-                this.setState({
-                    budgetList: response
-                });
-                this.setState({
-                    expense: Object.assign(this.state.expense, { budgetId: response[0].id })
-                })
-            });
-    }
+import { getCategories as getCategoriesAction } from '../../../actions/categoryActions';
+
+import {
+    updateTransactionField as updateTransactionFieldAction,
+    resetTransactionField as resetTransactionFieldAction,
+    addTransaction
+} from '../../../actions/transactionActions';
+
+
+class AddExpense extends Component {
 
     componentDidMount() {
-        this.fetchExpenseBudget();
+        this.props.getCategories();
     }
 
-    updateProperty(property, value) {
-        let obj = {};
-        obj[property] = value;
-        this.setState({
-            expense: Object.assign(this.state.expense, obj)
-        });
+    updateAmount = (amount) => {
+        this.props.updateTransactionField('amount', amount);
     }
 
-    addNewExpense() {
-        let { date, budgetId, desc, amount } = this.state.expense;
-        transactions.add({
-            year: DateTime.fromFormat(date, 'yyyy-LL-dd').year,
-            month: DateTime.fromFormat(date, 'yyyy-LL-dd').month,
-            date,
-            budgetId,
-            desc,
-            amount
-        })
-        .then((id)=>{
-            console.log(`[success] add expense (${id})`);
-            this.goBack();
-        });
+    onSetAmount = () => {
+        this.props.history.push('/expense/category');
     }
 
-    goBack() {
-        this.props.history.goBack();
+    onSelectCategory = (categoryId) => {
+        this.props.updateTransactionField('categoryId', categoryId);
+        this.props.history.push('/expense/desc');
+    }
+
+    onSave = async() => {
+        await this.props.saveTransation(this.props.transaction);
+        this.props.history.go(-3);
     }
 
     render() {
         return (
-            <div className="container home">
-                <h5 className="mt-3 mb-3">Add Expense</h5>
-                <form>
-                    <div className="form-group">
-                        <input type="text" className="form-control" placeholder="Description" required
-                            value={this.state.expense.desc}
-                            onChange={(e) => { this.updateProperty('desc', e.target.value) }} />
-                    </div>
-                    <div className="form-group">
-                        <input type="number" className="form-control" placeholder="Amount" required
-                            value={this.state.expense.amount}
-                            onChange={(e) => { this.updateProperty('amount', parseInt(e.target.value, 10)) }} />
-                    </div>
-                    <div className="form-group">
-                        <select className="form-control" value={this.state.expense.budgetId}
-                            onChange={(e) => { this.updateProperty('budgetId', parseInt(e.target.value, 10)) }}>
-                            {
-                                this.state.budgetList.map((b) => {
-                                    return <option value={b.id} key={b.id}>{b.desc}</option>
-                                })
-                            }
-                        </select>
-                    </div>
-                    <div className="form-group">
-                        <input type="date" className="form-control" placeholder="Date" required
-                            value={this.state.expense.date}
-                            onChange={(e) => { this.updateProperty('date', e.target.value) }} />
-                    </div>
-                    <div className="d-flex justify-content-center">
-                        <button type="button" className="btn btn-outline-primary mr-3 btn-block"
-                            onClick={this.goBack}>Cancel</button>
-                        <button type="button" className="btn btn-primary btn-block mt-0"
-                        onClick={this.addNewExpense}>Save</button>
-                    </div>
-                </form>
+            <div className="add-income-wrapper">
+                {
+                    (() => {
+                        let { step } = this.props.match.params;
+                        let { transaction, categories, updateTransactionField } = this.props;
+
+                        switch (step) {
+                            case 'amount':
+                                return <Amount onSetAmount={this.onSetAmount}
+                                    goBack={this.props.history.goBack}
+                                    transaction={transaction} updateAmount={this.updateAmount}>
+                                </Amount>;
+                            case 'category':
+                                categories = categories.filter(category => category.type === 'EXPENSE');
+                                return <Category onSelectCategory={this.onSelectCategory}
+                                    categories={categories}
+                                    goBack={this.props.history.goBack}
+                                    transaction={transaction}>
+                                </Category>;
+                            case 'desc':
+                                let selectedCaterogy = categories.find(category => category.id === transaction.categoryId);
+                                return <Description goBack={this.props.history.goBack}
+                                    onSave={this.onSave}
+                                    transaction={transaction}
+                                    category={selectedCaterogy ? selectedCaterogy.name : ''}
+                                    updateProperty={updateTransactionField}>
+                                </Description>;
+                            default:
+                                return '';
+                        }
+                    })()
+                }
             </div>
+
         );
     }
 }
 
-export default AddExpense;
+const mapStateToProp = state => ({
+    transaction: state.transaction,
+    categories: state.categories
+});
+
+const mapDispatchToProp = dispatch => ({
+    updateTransactionField(field, value) {
+        if (field === 'date') {
+            dispatch(updateTransactionFieldAction('year', DateTime.fromISO(value).year));
+            dispatch(updateTransactionFieldAction('month', DateTime.fromISO(value).month));
+        }
+        dispatch(updateTransactionFieldAction(field, value));
+    },
+    resetTransactionField() {
+        dispatch(resetTransactionFieldAction());
+    },
+    async getCategories() {
+        dispatch(await getCategoriesAction());
+    },
+    async saveTransation(transaction) {
+        await addTransaction(transaction);
+    }
+})
+
+export default withRouter(connect(mapStateToProp, mapDispatchToProp)(AddExpense));
