@@ -1,19 +1,18 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
+import { Provider } from 'react-redux';
+import store from './store';
 import './index.scss';
 import App from './App';
-
-import 'bootstrap/dist/css/bootstrap.css';
-import 'jquery';
-import 'bootstrap/dist/js/bootstrap';
-import 'font-awesome/css/font-awesome.css';
 
 import db from './db/db';
 db.initialize();
 
 ReactDOM.render((
-    <Router>
-        <App />
-    </Router>
+    <Provider store={store}>
+        <Router>
+            <App />
+        </Router>
+    </Provider>
 ), document.getElementById('root'));
